backend/requirements.txt
docx2txt==0.8
fastapi==0.105.0
langchain==0.0.352
milvus==2.3.3
openai==1.6.1
pandas==2.1.4
pymilvus==2.3.4
python-dotenv==1.0.0
sentence_transformers==2.2.2
torch==2.1.2
towhee==1.1.3
towhee.models==1.1.3
transformers==4.36.2
unstructured==0.11.6
uvicorn==0.25.0
psycopg2==2.9.9
